+++ docs/changelog.md
@@ -6,6 +6,8 @@ List of changes made to waiter.
 
 - Added `httr_progress` function to use instead of 
 `httr::progress`.
+- Added `withProgressWaitress`, `setProgressWaitress`, and
+`incProgressWairess` analogous to the same functions in shiny.
 
 ## waiter 0.2.3